Redirect issue fix
