+++ src/components/widgets/megaMenu/GuiFields.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import WidgetBuilder from '@src/routes/(app)/builder/WidgetBuilder.svelte';
+	import WidgetBuilder from '@components/system/builder/WidgetBuilder.svelte';
 
 	export let value: Array<Array<any>> = [];
 	//console.log(value);
