+++ src/routes/(app)/collection/[...collectionName]/components/CollectionPermission.svelte
@@ -1,0 +1,13 @@
+<script lang="ts">
+	import Permissions from '@src/components/Permissions.svelte';
+	import { tabSet } from '@stores/store';
+</script>
+
+<!-- Permissions -->
+<Permissions />
+
+<!-- Buttons -->
+<div class="flex justify-between">
+	<button type="button" on:click={() => ($tabSet = 0)} class="variant-filled-secondary btn mt-2 justify-end">Previous</button>
+	<button type="button" on:click={() => ($tabSet = 2)} class="variant-filled-tertiary btn mt-2 dark:variant-filled-primary">Next</button>
+</div>
